Continuing refactoring
