+++ dockstore-client/Hello.java
@@ -1,0 +1,1 @@
+public class Hello {}
