+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/GitLabSourceCodeRepo.java
@@ -21,10 +21,10 @@ import java.net.URLEncoder;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.google.common.base.Optional;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -35,6 +35,7 @@ import io.dockstore.webservice.core.Entry;
 import io.dockstore.webservice.core.SourceFile;
 import io.dockstore.webservice.core.Workflow;
 import io.dockstore.webservice.core.WorkflowVersion;
+import io.dockstore.webservice.languages.LanguageHandlerFactory;
 import io.dockstore.webservice.resources.ResourceUtilities;
 import org.apache.http.HttpStatus;
 import org.apache.http.client.HttpClient;
@@ -199,25 +200,7 @@ public class GitLabSourceCodeRepo extends SourceCodeRepoInterface {
                     }
 
                     // Use default test parameter file if either new version or existing version that hasn't been edited
-                    if (!version.isDirtyBit() && workflow.getDefaultTestParameterFilePath() != null) {
-                        // Set Filetype
-                        SourceFile.FileType testJsonType = null;
-                        if (identifiedType.equals(SourceFile.FileType.DOCKSTORE_CWL)) {
-                            testJsonType = SourceFile.FileType.CWL_TEST_JSON;
-                        } else if (identifiedType.equals(SourceFile.FileType.DOCKSTORE_WDL)) {
-                            testJsonType = SourceFile.FileType.WDL_TEST_JSON;
-                        }
-
-                        // Check if test parameter file has already been added
-                        final SourceFile.FileType finalFileType = testJsonType;
-                        long duplicateCount = version.getSourceFiles().stream().filter((SourceFile v) -> v.getPath().equals(workflow.getDefaultTestParameterFilePath()) && v.getType() == finalFileType).count();
-                        if (duplicateCount == 0) {
-                            SourceFile testJsonSourceFile = getSourceFile(workflow.getDefaultTestParameterFilePath(), id, branchName, testJsonType);
-                            if (testJsonSourceFile != null) {
-                                version.getSourceFiles().add(testJsonSourceFile);
-                            }
-                        }
-                    }
+                    createTestParameterFiles(workflow, id, branchName, version, identifiedType);
 
                     workflow.addWorkflowVersion(
                             combineVersionAndSourcefile(sourceFile, workflow, identifiedType, version, existingDefaults));
@@ -227,6 +210,8 @@ public class GitLabSourceCodeRepo extends SourceCodeRepoInterface {
 
         return workflow;
     }
+
+
 
     @Override
     public String getRepositoryId(Entry entry) {
@@ -318,7 +303,8 @@ public class GitLabSourceCodeRepo extends SourceCodeRepoInterface {
      * @param type
      * @return source file
      */
-    private SourceFile getSourceFile(String path, String id, String branch, SourceFile.FileType type) {
+    @Override
+    public SourceFile getSourceFile(String path, String id, String branch, SourceFile.FileType type) {
         // TODO: should we even be creating a sourcefile before checking that it is valid?
         // I think it is fine since in the next part we just check that source file has content or not (no content is like null)
         SourceFile file = null;
@@ -329,14 +315,7 @@ public class GitLabSourceCodeRepo extends SourceCodeRepoInterface {
             boolean validWorkflow;
             file = new SourceFile();
 
-            if (type == SourceFile.FileType.DOCKSTORE_CWL) {
-                validWorkflow = checkValidCWLWorkflow(content);
-            } else if (type == SourceFile.FileType.DOCKSTORE_WDL) {
-                validWorkflow = checkValidWDLWorkflow(content);
-            } else {
-                // Must be a testjson file
-                validWorkflow = true;
-            }
+            validWorkflow = LanguageHandlerFactory.getInterface(type, this).isValidWorkflow(content);
 
             if (validWorkflow) {
                 file.setType(type);
