+++ dockstore-webservice/src/main/resources/swagger.yaml
@@ -3908,7 +3908,7 @@ definitions:
         - "WDL_TEST_JSON"
         - "NEXTFLOW"
         - "NEXTFLOW_CONFIG"
-        - "NEXTFLOW_PARAMS"
+        - "NEXTFLOW_TEST_PARAMS"
       content:
         type: "string"
         description: "Cache for the contents of the target file"
